Use a QueryBuilder instead of select directly
The main problem with using select directly was type inference.
We need to provide the type of the Model in order to get good inference
on the keys we're using in select. But we want to not manually provide
the key type, since it needs to be repeated by passing the same strings
to select anyways. By providing a QueryBuilder type that we then call
select on, we can just provide the first type parameter when calling
`query` and have the key type be inferred.